Upload cache hidden field should always be present
1a67a90 refactored a common form pattern to a builder object, but the
semantics of the form changed slightly.

This causes the file upload data attributes to be empty, which removes
the entry entirely from the params. Our uploading controller code
doesn't validate uploaded data if it isn't present in the params at
all, which caused ImageData (and other) records to be created with NULL
ImageData, which crashes the display of those editions in the admin.

Restoring the original semantics fixes the issue.